pyproject.toml
[tool.poetry]
name = "ekg-arrhythmia-ischemia-classifier"
version = "0.1.0"
description = "A classifier for detecting arrhythmia and ischemia from EKG data."
authors = ["Alejandro Garcia", "Dev Kansagra", "Dylan Figueredo", "Sannan Khalid"]
license = "MIT"
readme = "README.md"

[[tool.poetry.packages]]
include = "ekgclf"
from = "src"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.1"
torchaudio = "^2.3.1"
numpy = "^2.0.4"
scipy = "^1.13.1"
pandas = "^2.2.2"
wfdb = "^4.1.2"
scikit-learn = "^1.5.1"
hydra-core = "^1.3.2"
pydantic = "^2.8.2"
mlflow = "^2.15.1"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.5", extras = ["standard"] }
matplotlib = "^3.9.1"
tqdm = "^4.66.4"
pyyaml = "^6.0.2"
prometheus-client = "^0.20.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.6.9"
black = "^24.8.0"
isort = "^5.13.2"
nbstripout = "^0.7.1"
requests = "^2.32.3"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"